move app to deprecated
