+++ packages/promise-cache/src/index.ts
@@ -1,4 +1,4 @@
 export { useApi, createApi } from "./lib/api";
 export { useCache, AppProvider } from "./lib/context";
-export { api, useApp, createApp } from "./lib/application";
 export { Hydration, SuspenseBoundary } from "./lib/Hydration";
+export { api, useApp, createApp } from "./deprecated/application";
